requirements.txt
Django>=6.0,<6.1
gunicorn>=23.0.0,<24.0.0
whitenoise>=6.9.0,<7.0.0
django-summernote>=0.8.20.0,<0.9.0
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.13.0,<5.0.0
Pillow>=11.0.0,<12.0.0
django-dbbackup>=4.2.0,<5.0.0
rapidocr-onnxruntime>=1.4.4,<2.0.0

schedule>=1.2.0,<2.0.0

schedule>=1.2.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
